[swag] warning

+++ src/CommandLineParser.cpp
@@ -139,15 +139,15 @@ void CommandLineParser::logArguments(const Utf8& cmd)
 {
     Utf8 line0, line1;
 
-    int columns[10] = {4};
+    size_t columns[10] = {4};
 
     static const size_t SPACE = 4;
 
-    columns[0] = SPACE + (int) strlen("Argument");
-    columns[1] = SPACE + (int) strlen("Short");
-    columns[2] = SPACE + (int) strlen("Value");
-    columns[3] = SPACE + (int) strlen("Default");
-    columns[4] = SPACE + (int) strlen("Help");
+    columns[0] = SPACE + strlen("Argument");
+    columns[1] = SPACE + strlen("Short");
+    columns[2] = SPACE + strlen("Value");
+    columns[3] = SPACE + strlen("Default");
+    columns[4] = SPACE + strlen("Help");
 
     for (auto arg : longNameArgs)
     {
@@ -163,10 +163,10 @@ void CommandLineParser::logArguments(const Utf8& cmd)
         columns[2] = max(columns[2], value.length() + SPACE);
         columns[3] = max(columns[3], defaultVal.length() + SPACE);
 
-        columns[4] = max(columns[4], (int) strlen(oneArg->help) + SPACE);
+        columns[4] = max(columns[4], strlen(oneArg->help) + SPACE);
     }
 
-    int total = 0;
+    size_t total = 0;
 
     line0 = "Argument";
     line1 = "--------";
